Use decode/encode double instead of string for `BigDecimalSerializer`
A JSON number is represented as an unquoted value and should perfectly fit a double according to the JSON spec. Unfortunatly, we cannot directly use `Double` because of weird templating issues of the JSON schema code generator, but this at least encodes and decodes the values as a number and not as a string - which actually caused crashes when loading several CSAF documents in the wild.